Explicitly truncate the result

+++ NULLC/Executor.cpp
@@ -1920,7 +1920,7 @@ void Executor::Run(unsigned int functionID, const char *arguments)
 				{
 					unsigned length = *(int*)(((char*)genStackPtr) + sizeof(void*));
 					char *copy = (char*)NULLC::AllocObject(exLinker->exTypes[type.subType].size * length);
-					memcpy(copy, vmLoadPointer(genStackPtr), exLinker->exTypes[type.subType].size * length);
+					memcpy(copy, vmLoadPointer(genStackPtr), unsigned(exLinker->exTypes[type.subType].size * length));
 					vmStorePointer(genStackPtr, copy);
 				}else{
 					unsigned int objSize = type.size;
